+++ libs/langchain-mongodb/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ---
 
+## Changes in version 0.12.0 (XXXX/XX/XX)
+
+- Add support for auto-embeddings in retrievers:
+
+  - `MongoDBAtlasHybridSearchRetriever`
+  - `MongoDBAtlasParentDocumentRetriever`
+  - `MongoDBAtlasSelfQueryRetriever`
+
 ## Changes in version 0.11.0 (2026/01/15)
 
 - Add support for auto-embeddings in `MongoDBAtlasVectorSearch`.